Relocate user group methods from permissions manager to user manager. Continue renaming PermissionsManager to EntityPermissionsManager
